__app__/requirements.txt
azure-functions
beautifulsoup4
requests